backend/requirements.txt

fastapi==0.115.2
uvicorn[standard]==0.30.1
requests==2.32.3
beautifulsoup4==4.12.3
trafilatura==1.7.0
playwright==1.47.0
python-dotenv==1.0.1
tiktoken==0.7.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0
chromadb==0.5.5
openai==1.55.0
jinja2==3.1.4
